+++ frontend/src/layout/Sidebar.js
@@ -31,7 +31,6 @@ import { useEscapeKey } from 'lib/hooks/useEscapeKey'
 import { ToolbarModal } from '~/layout/ToolbarModal/ToolbarModal'
 import whiteLogo from './../../public/posthog-logo-white.svg'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { useCommands } from 'lib/components/CommandPalette/commandLogic'
 
 const itemStyle = { display: 'flex', alignItems: 'center' }
 
@@ -94,42 +93,42 @@ export function Sidebar({ user, sidebarCollapsed, setSidebarCollapsed }) {
     }
 
     /* Command palette custom commands */
-    const commandBilling = {
-        key: 'billing',
-        prefixes: [],
-        resolver: () => {
-            return [
-                {
-                    key: 'billing',
-                    icon: WalletOutlined,
-                    display: 'go to billing',
-                    executor: ({ push }) => {
-                        push('/billing')
-                    },
-                },
-            ]
-        },
-    }
+    // const commandBilling = {
+    //     key: 'billing',
+    //     prefixes: [],
+    //     resolver: () => {
+    //         return [
+    //             {
+    //                 key: 'billing',
+    //                 icon: WalletOutlined,
+    //                 display: 'go to billing',
+    //                 executor: ({ push }) => {
+    //                     push('/billing')
+    //                 },
+    //             },
+    //         ]
+    //     },
+    // }
 
-    const commandLicense = {
-        key: 'licenses',
-        prefixes: [],
-        resolver: () => {
-            return [
-                {
-                    key: 'license',
-                    icon: LockOutlined,
-                    display: 'go to licenses',
-                    executor: ({ push }) => {
-                        push('/setup/licenses')
-                    },
-                },
-            ]
-        },
-    }
+    // const commandLicense = {
+    //     key: 'licenses',
+    //     prefixes: [],
+    //     resolver: () => {
+    //         return [
+    //             {
+    //                 key: 'license',
+    //                 icon: LockOutlined,
+    //                 display: 'go to licenses',
+    //                 executor: ({ push }) => {
+    //                     push('/setup/licenses')
+    //                 },
+    //             },
+    //         ]
+    //     },
+    // }
 
-    useCommands([commandBilling], featureFlags['billing-management-page'])
-    useCommands([commandLicense], !user.is_multi_tenancy && user.ee_available)
+    // useCommands([commandBilling], featureFlags['billing-management-page'])
+    // useCommands([commandLicense], !user.is_multi_tenancy && user.ee_available)
     /* End command palette custom commands */
 
     return (
