+++ frontend/src/lib/components/CommandPalette/index.tsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useState, useRef } from 'react'
 import { useOutsideClickHandler } from 'lib/utils'
 import { useHotkeys } from 'react-hotkeys-hook'
-import { useValues } from 'kea'
-import { useCommandsSearch, CommandResult as CommandResultType } from './commandLogic'
+import { useMountedLogic, useValues, useActions } from 'kea'
+import { CommandResult as CommandResultType, commandLogic } from './commandLogic'
 import { CommandSearch } from './CommandSearch'
 import { CommandResult } from './CommandResult'
-import { GlobalCommands } from './GlobalCommands'
 import styled from 'styled-components'
 import { userLogic } from 'scenes/userLogic'
 import { useEventListener } from 'lib/hooks/useEventListener'
@@ -69,9 +68,13 @@ const PaletteError = styled.div`
 `
 
 export function CommandPalette(): JSX.Element | null {
+    useMountedLogic(commandLogic)
+
+    const { setSearchInput: setInput } = useActions(commandLogic)
+    const { searchInput: input, commandSearchResults } = useValues(commandLogic)
     const boxRef = useRef<HTMLDivElement | null>(null)
     const [state] = useState({ error: null, title: null })
-    const [input, setInput] = useState('')
+
     const [isPaletteShown, setIsPaletteShown] = useState(false)
     const { user } = useValues(userLogic)
 
@@ -106,19 +109,17 @@ export function CommandPalette(): JSX.Element | null {
     }, [isPaletteShown])
 
     useEffect(() => {
-        if (Object.keys(commandsSearch(input)).length - 1 > activeResultIndex) {
+        if (commandSearchResults.length - 1 > activeResultIndex) {
             setActiveResultIndex(0)
         }
     }, [input])
-
-    const commandsSearch = useCommandsSearch()
 
     const _handleKeyDown = useCallback(
         (e: KeyboardEvent) => {
             if (isPaletteShown) {
                 if (e.key === 'ArrowDown') {
                     setActiveResultIndex((prevIndex) => {
-                        if (prevIndex === Object.keys(commandsSearch(input)).length - 1) return prevIndex
+                        if (prevIndex === commandSearchResults.length - 1) return prevIndex
                         else return prevIndex + 1
                     })
                 } else if (e.key === 'ArrowUp') {
@@ -137,7 +138,7 @@ export function CommandPalette(): JSX.Element | null {
     const _handleEnterDown = useCallback(
         (e: KeyboardEvent) => {
             if (e.key === 'Enter') {
-                handleCommandSelection(commandsSearch(input)[activeResultIndex])
+                handleCommandSelection(commandSearchResults[activeResultIndex])
             }
         },
         [activeResultIndex, input]
@@ -147,7 +148,6 @@ export function CommandPalette(): JSX.Element | null {
 
     return (
         <>
-            <GlobalCommands />
             {!user || !isPaletteShown ? null : (
                 <CommandPaletteContainer>
                     <CommandPaletteBox ref={boxRef} className="bg-dark">
@@ -161,14 +161,13 @@ export function CommandPalette(): JSX.Element | null {
                         />
                         {state.error && <PaletteError>{state.error}</PaletteError>}
                         <ResultsContainer>
-                            {commandsSearch(input).map((result, index) => (
+                            {commandSearchResults.map((result, index) => (
                                 <CommandResult
                                     focused={activeResultIndex === index}
                                     key={`command-result-${index}`}
                                     result={result}
                                     handleSelection={handleCommandSelection}
                                     onMouseOver={(): void => setActiveResultIndex(-1)}
-                                    onMouseOut={(): void => setActiveResultIndex(0)}
                                 />
                             ))}
                         </ResultsContainer>
