+++ frontend/src/scenes/insights/InsightTabs/FunnelTab/FunnelTab.tsx
@@ -1,17 +1,19 @@
 import React from 'react'
-import { useValues, useActions } from 'kea'
+import { useValues, useActions, useMountedLogic } from 'kea'
 import { PropertyFilters } from 'lib/components/PropertyFilters/PropertyFilters'
 
 import { funnelLogic } from 'scenes/funnels/funnelLogic'
 import { actionsModel } from '~/models/actionsModel'
 import { userLogic } from 'scenes/userLogic'
-import { ActionFilter } from '../ActionFilter/ActionFilter'
+import { ActionFilter } from '../../ActionFilter/ActionFilter'
 import { Link } from 'lib/components/Link'
 import { Button, Row } from 'antd'
 import { useState } from 'react'
-import SaveModal from '../SaveModal'
+import SaveModal from '../../SaveModal'
+import { funnelCommandLogic } from './funnelCommandLogic'
 
 export function FunnelTab(): JSX.Element {
+    useMountedLogic(funnelCommandLogic)
     const { isStepsEmpty, filters, stepsWithCount } = useValues(funnelLogic)
     const { loadFunnel, clearFunnel, setFilters, saveFunnelInsight } = useActions(funnelLogic)
     const { actions, actionsLoading } = useValues(actionsModel)
