+++ frontend/src/scenes/insights/InsightTabs/index.ts
@@ -2,4 +2,4 @@ export * from './RetentionTab'
 export * from './SessionTab'
 export * from './TrendTab'
 export * from './PathTab'
-export * from './FunnelTab'
+export * from './FunnelTab/FunnelTab'
